# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PublicTag": {
				"description": "前台公開標籤物件（不含後台管理欄位）",
				"properties": {
					"id": {
						"type": "string",
						"description": "標籤唯一識別碼（UUID）",
						"example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
					},
					"name": {
						"type": "string",
						"description": "標籤名稱",
						"example": "TypeScript"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PublicTag-Array_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/PublicTag"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PublicTag_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PublicTag"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PublicPost": {
				"description": "前台公開文章物件（不含後台管理欄位）",
				"properties": {
					"id": {
						"type": "string",
						"description": "文章唯一識別碼（UUID）",
						"example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
					},
					"title": {
						"type": "string",
						"description": "文章標題",
						"example": "我的第一篇文章"
					},
					"content": {
						"type": "string",
						"description": "文章內文",
						"example": "這是文章內容..."
					},
					"author": {
						"type": "string",
						"description": "作者名稱",
						"example": "Alice"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "建立時間（UTC）",
						"example": "2024-01-15T08:30:00.000Z"
					}
				},
				"required": [
					"id",
					"title",
					"content",
					"author",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PublicPost-Array_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/PublicPost"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PublicPost_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PublicPost"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Banner": {
				"properties": {
					"id": {
						"type": "string",
						"description": "唯一識別碼（UUID，由後端自動產生）",
						"example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
					},
					"title": {
						"type": "string",
						"description": "banner 名稱",
						"example": "logo+文字"
					},
					"type": {
						"type": "string",
						"description": "banner 類型",
						"example": "圖文|圖"
					},
					"img": {
						"type": "string",
						"description": "圖片網址"
					},
					"isActive": {
						"type": "boolean",
						"description": "是否啟用"
					},
					"content": {
						"type": "string",
						"description": "內容"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "建立時間（UTC）",
						"example": "2024-01-15T08:30:00.000Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "最後更新時間（UTC），未更新過則為 null",
						"example": "2024-01-20T12:00:00.000Z"
					}
				},
				"required": [
					"id",
					"title",
					"type",
					"img",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Banner-or-undefined_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Banner"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Tag": {
				"description": "後台完整標籤物件",
				"properties": {
					"id": {
						"type": "string",
						"description": "標籤唯一識別碼（UUID，由後端自動產生）",
						"example": "b2c3d4e5-f6a7-8901-bcde-f12345678901"
					},
					"name": {
						"type": "string",
						"description": "標籤名稱",
						"example": "TypeScript"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "建立時間（UTC）",
						"example": "2024-01-15T08:30:00.000Z"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Tag-Array_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_Tag_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/Tag"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTagDto": {
				"description": "新增標籤時的請求資料",
				"properties": {
					"name": {
						"type": "string",
						"description": "標籤名稱（必填）",
						"example": "TypeScript"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTagDto": {
				"description": "更新標籤時的請求資料",
				"properties": {
					"name": {
						"type": "string",
						"description": "新的標籤名稱",
						"example": "JavaScript"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostModel": {
				"description": "後台完整文章物件（從資料庫/記憶體讀出來的完整資料）",
				"properties": {
					"id": {
						"type": "string",
						"description": "文章唯一識別碼（UUID，由後端自動產生）",
						"example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
					},
					"title": {
						"type": "string",
						"description": "文章標題",
						"example": "我的第一篇文章"
					},
					"content": {
						"type": "string",
						"description": "文章內文（支援純文字或 Markdown）",
						"example": "這是文章內容..."
					},
					"author": {
						"type": "string",
						"description": "作者名稱",
						"example": "Alice"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "建立時間（UTC）",
						"example": "2024-01-15T08:30:00.000Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "最後更新時間（UTC），未更新過則為 null",
						"example": "2024-01-20T12:00:00.000Z"
					}
				},
				"required": [
					"id",
					"title",
					"content",
					"author",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PostModel-Array_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/PostModel"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PostModel_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/PostModel"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostDto": {
				"description": "新增文章時的請求資料（id 和時間戳由後端自動產生）",
				"properties": {
					"title": {
						"type": "string",
						"description": "文章標題（必填）",
						"example": "我的第一篇文章"
					},
					"content": {
						"type": "string",
						"description": "文章內文（必填）",
						"example": "這是文章內容..."
					},
					"author": {
						"type": "string",
						"description": "作者名稱（必填）",
						"example": "Alice"
					}
				},
				"required": [
					"title",
					"content",
					"author"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePostDto": {
				"description": "更新文章時的請求資料（所有欄位皆為選填，只傳要改的欄位即可）",
				"properties": {
					"title": {
						"type": "string",
						"description": "新的文章標題",
						"example": "更新後的標題"
					},
					"content": {
						"type": "string",
						"description": "新的文章內文",
						"example": "更新後的內容..."
					},
					"author": {
						"type": "string",
						"description": "新的作者名稱",
						"example": "Bob"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"BannerModel": {
				"properties": {
					"id": {
						"type": "string",
						"description": "唯一識別碼（UUID，由後端自動產生）",
						"example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
					},
					"title": {
						"type": "string",
						"description": "banner 名稱",
						"example": "logo+文字"
					},
					"type": {
						"type": "string",
						"description": "banner 類型",
						"example": "圖文|圖"
					},
					"img": {
						"type": "string",
						"description": "圖片網址"
					},
					"isActive": {
						"type": "boolean",
						"description": "是否啟用"
					},
					"content": {
						"type": "string",
						"description": "內容"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "建立時間（UTC）",
						"example": "2024-01-15T08:30:00.000Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "最後更新時間（UTC），未更新過則為 null",
						"example": "2024-01-20T12:00:00.000Z"
					}
				},
				"required": [
					"id",
					"title",
					"type",
					"img",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_BannerModel-Array_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/BannerModel"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_BannerModel_": {
				"description": "統一 API 回傳格式",
				"properties": {
					"data": {
						"$ref": "#/components/schemas/BannerModel"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBannerDto": {
				"properties": {
					"title": {
						"type": "string",
						"description": "banner 名稱",
						"example": "logo+文字"
					},
					"type": {
						"type": "string",
						"description": "banner 類型",
						"example": "圖文|圖"
					},
					"img": {
						"type": "string",
						"description": "圖片網址"
					},
					"content": {
						"type": "string",
						"description": "內容"
					}
				},
				"required": [
					"title",
					"type",
					"img"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBannerDto": {
				"properties": {
					"title": {
						"type": "string",
						"description": "banner 名稱",
						"example": "logo+文字"
					},
					"type": {
						"type": "string",
						"description": "banner 類型",
						"example": "圖文|圖"
					},
					"img": {
						"type": "string",
						"description": "圖片網址"
					},
					"content": {
						"type": "string",
						"description": "內容"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "angular-blog-server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/public/tags": {
			"get": {
				"operationId": "GetTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PublicTag-Array_"
								}
							}
						}
					}
				},
				"description": "取得所有標籤清單",
				"tags": [
					"Public - Tags"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/public/tags/{id}": {
			"get": {
				"operationId": "GetTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PublicTag_"
								}
							}
						}
					},
					"404": {
						"description": "Tag not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "根據 ID 取得單一標籤",
				"tags": [
					"Public - Tags"
				],
				"security": [],
				"parameters": [
					{
						"description": "標籤 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/public/posts": {
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PublicPost-Array_"
								}
							}
						}
					}
				},
				"description": "取得所有文章清單",
				"tags": [
					"Public - Posts"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/public/posts/{id}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PublicPost_"
								}
							}
						}
					},
					"404": {
						"description": "Post not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "根據 ID 取得單篇文章",
				"tags": [
					"Public - Posts"
				],
				"security": [],
				"parameters": [
					{
						"description": "文章 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/public/banner": {
			"get": {
				"operationId": "GetPublicBanner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Banner-or-undefined_"
								}
							}
						}
					}
				},
				"description": "取得前台 banner",
				"tags": [
					"Public - Banner"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/admin/tags": {
			"get": {
				"operationId": "GetTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tag-Array_"
								}
							}
						}
					}
				},
				"description": "取得所有標籤清單（含後台管理欄位）",
				"tags": [
					"Admin - Tags"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTag",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tag_"
								}
							}
						}
					}
				},
				"description": "新增一個標籤",
				"tags": [
					"Admin - Tags"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTagDto"
							}
						}
					}
				}
			}
		},
		"/api/admin/tags/{id}": {
			"get": {
				"operationId": "GetTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tag_"
								}
							}
						}
					},
					"404": {
						"description": "Tag not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "根據 ID 取得單一標籤（含後台管理欄位）",
				"tags": [
					"Admin - Tags"
				],
				"security": [],
				"parameters": [
					{
						"description": "標籤 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Tag_"
								}
							}
						}
					},
					"404": {
						"description": "Tag not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "更新指定標籤名稱",
				"tags": [
					"Admin - Tags"
				],
				"security": [],
				"parameters": [
					{
						"description": "標籤 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTagDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTag",
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"404": {
						"description": "Tag not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "刪除指定標籤",
				"tags": [
					"Admin - Tags"
				],
				"security": [],
				"parameters": [
					{
						"description": "標籤 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/posts": {
			"get": {
				"operationId": "GetPosts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PostModel-Array_"
								}
							}
						}
					}
				},
				"description": "取得所有文章清單（含後台管理欄位）",
				"tags": [
					"Admin - Posts"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PostModel_"
								}
							}
						}
					}
				},
				"description": "新增一篇文章",
				"tags": [
					"Admin - Posts"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostDto"
							}
						}
					}
				}
			}
		},
		"/api/admin/posts/{id}": {
			"get": {
				"operationId": "GetPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PostModel_"
								}
							}
						}
					},
					"404": {
						"description": "Post not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "根據 ID 取得單篇文章（含後台管理欄位）",
				"tags": [
					"Admin - Posts"
				],
				"security": [],
				"parameters": [
					{
						"description": "文章 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PostModel_"
								}
							}
						}
					},
					"404": {
						"description": "Post not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "更新指定文章的內容（只傳要修改的欄位即可）",
				"tags": [
					"Admin - Posts"
				],
				"security": [],
				"parameters": [
					{
						"description": "文章 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePost",
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"404": {
						"description": "Post not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "刪除指定文章",
				"tags": [
					"Admin - Posts"
				],
				"security": [],
				"parameters": [
					{
						"description": "文章 ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/admin/banner": {
			"get": {
				"operationId": "GetBanners",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BannerModel-Array_"
								}
							}
						}
					}
				},
				"description": "取得所有後台 banner",
				"tags": [
					"Admin - Banner"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateBanner",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BannerModel_"
								}
							}
						}
					}
				},
				"description": "新增後台 banner",
				"tags": [
					"Admin - Banner"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBannerDto"
							}
						}
					}
				}
			}
		},
		"/api/admin/banner/{id}": {
			"get": {
				"operationId": "GetBanner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BannerModel_"
								}
							}
						}
					},
					"404": {
						"description": "Banner not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "取得單一後台 banner",
				"tags": [
					"Admin - Banner"
				],
				"security": [],
				"parameters": [
					{
						"description": "banner ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBanner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_BannerModel_"
								}
							}
						}
					},
					"404": {
						"description": "Banner not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "更新指定banner的內容（只傳要修改的欄位即可）",
				"tags": [
					"Admin - Banner"
				],
				"security": [],
				"parameters": [
					{
						"description": "banner ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBannerDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBanner",
				"responses": {
					"204": {
						"description": "Deleted"
					},
					"404": {
						"description": "Banner not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "刪除指定banner",
				"tags": [
					"Admin - Banner"
				],
				"security": [],
				"parameters": [
					{
						"description": "banner ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}